% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphafoldConfidence_module_ui.R
\name{alphafoldConfidenceOutputUI}
\alias{alphafoldConfidenceOutputUI}
\title{Output UI components for the alphafoldConfidence module}
\usage{
alphafoldConfidenceOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; when \code{TRUE} (the default) the plot output can be resized by dragging.}
}
\value{
A Shiny plotlyOutput for the confidence plot.
}
\description{
Output UI components for the alphafoldConfidence module
}
\examples{
alphafoldConfidenceOutputUI("af")
}
\author{
Jared Andrews
}
