% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crisprScreenRank_module_ui.R
\name{crisprScreenRankOutputUI}
\alias{crisprScreenRankOutputUI}
\title{Output UI components for the crisprScreenRank module}
\usage{
crisprScreenRankOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical, whether to wrap the output in a resizable container.}
}
\value{
A Shiny plotlyOutput for the rank plot.
}
\description{
Output UI components for the crisprScreenRank module
}
\examples{
crisprScreenRankOutputUI("crispr")
}
\author{
Jared Andrews
}
