% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoFreqPlot_module_ui.R
\name{dittoFreqPlotOutputUI}
\alias{dittoFreqPlotOutputUI}
\title{Output UI components for the dittoFreqPlot module}
\usage{
dittoFreqPlotOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; when \code{TRUE} (the default) the plot output is
wrapped in \code{\link[shinyjqui:jqui_resizable]{shinyjqui::jqui_resizable()}} so it can be resized by dragging.}
}
\value{
A Shiny plotlyOutput for the frequency plot.
}
\description{
This should be placed in the UI where the plot should be shown.
}
\author{
Jacob Martin, Jared Andrews
}
