% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_cn_genes}
\alias{example_cn_genes}
\title{Example gene coordinates for copy number segment labeling (simulated)}
\format{
A \code{GRanges} with 6 ranges and 1 metadata column:
\describe{
\item{gene_name}{Example gene symbol}
}
}
\source{
Simulated in \code{data-raw/generate_example_cn_segment.R}; no real
biological data are included.
}
\usage{
example_cn_genes
}
\description{
A small \code{GRanges} of example gene coordinates overlapping the simulated
gain/loss regions in \link{example_cn_segment}, provided as ready-to-use example
data for gene labeling in the \code{\link[=cnSegmentPlotServer]{cnSegmentPlotServer()}} module and
\code{\link[=cnSegmentPlot]{cnSegmentPlot()}}.
}
\examples{
library(sciVizModules)
data(example_cn_genes)
example_cn_genes

}
\seealso{
\link{example_cn_segment}
}
\author{
Jacob Martin, Jared Andrews
}
\keyword{datasets}
