% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaEnrichmentPlot_module_ui.R
\name{gseaEnrichmentPlotOutputUI}
\alias{gseaEnrichmentPlotOutputUI}
\title{Output UI components for the gseaEnrichmentPlot module}
\usage{
gseaEnrichmentPlotOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; when \code{TRUE} (the default) the plot output can be resized by dragging.}
}
\value{
A Shiny plotlyOutput for the GSEA plot.
}
\description{
Output UI components for the gseaEnrichmentPlot module
}
\examples{
gseaEnrichmentPlotOutputUI("gsea")
}
\author{
Jared Andrews
}
