% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwasQQPlot_module_ui.R
\name{gwasQQPlotOutputUI}
\alias{gwasQQPlotOutputUI}
\title{Output UI components for the gwasQQPlot module}
\usage{
gwasQQPlotOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical, whether to wrap the output in a resizable container.}
}
\value{
A Shiny plotlyOutput for the QQ plot.
}
\description{
Output UI components for the gwasQQPlot module
}
\examples{
gwasQQPlotOutputUI("qq")
}
\author{
Jared Andrews
}
