% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphafold_readers.R
\name{.af_plddt_category}
\alias{.af_plddt_category}
\title{AlphaFold DB pLDDT confidence bands}
\usage{
.af_plddt_category(x)
}
\arguments{
\item{x}{Numeric pLDDT values.}
}
\value{
A factor with levels "Very high", "Confident", "Low", "Very low".
}
\description{
AlphaFold DB pLDDT confidence bands
}
\author{
Jared Andrews
}
\keyword{internal}
