% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphafold_readers.R
\name{.assert_alphafold}
\alias{.assert_alphafold}
\title{Check that an object is from read_alphafold()}
\usage{
.assert_alphafold(x, arg = "data")
}
\arguments{
\item{x}{The object to check.}

\item{arg}{Argument name for the error message.}
}
\value{
\code{x}, invisibly.
}
\description{
Check that an object is from read_alphafold()
}
\author{
Jared Andrews
}
\keyword{internal}
