% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ditto_helpers.R
\name{.assert_ditto_object}
\alias{.assert_ditto_object}
\title{Stop when an object is not dittoSeq-compatible}
\usage{
.assert_ditto_object(object, arg = "object")
}
\arguments{
\item{object}{An object to validate.}

\item{arg}{The argument name to report in the error message.}
}
\value{
Invisibly \code{TRUE} when valid; otherwise throws an error.
}
\description{
Stop when an object is not dittoSeq-compatible
}
\author{
Jacob Martin, Jared Andrews
}
\keyword{internal}
