% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_helpers.R
\name{.assert_gsea}
\alias{.assert_gsea}
\title{Validate a GSEA input for the module}
\usage{
.assert_gsea(x, arg = "data")
}
\arguments{
\item{x}{The input.}

\item{arg}{Argument name for error messages.}
}
\value{
The normalised bundle from \code{\link[=.gsea_input]{.gsea_input()}}.
}
\description{
Validate a GSEA input for the module
}
\author{
Jared Andrews
}
\keyword{internal}
