% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_wrapper_helpers.R
\name{.assert_heatmap_packages}
\alias{.assert_heatmap_packages}
\title{Stop unless the heatmap packages are installed}
\usage{
.assert_heatmap_packages()
}
\value{
\code{TRUE}, invisibly.
}
\description{
Stop unless the heatmap packages are installed
}
\author{
Jared Andrews
}
\keyword{internal}
