% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcatools_helpers.R
\name{.assert_pca}
\alias{.assert_pca}
\title{Check that an object is a PCAtools \code{pca} object}
\usage{
.assert_pca(x, arg = "data")
}
\arguments{
\item{x}{The object to check.}

\item{arg}{Argument name for the error message.}
}
\value{
\code{x}, invisibly.
}
\description{
The PCAtools modules read the object's fields directly (it is a plain list),
so this only checks the fields they use are present.
}
\author{
Jared Andrews
}
\keyword{internal}
