% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_helpers.R
\name{.assert_se}
\alias{.assert_se}
\title{Check that an object is a SummarizedExperiment}
\usage{
.assert_se(x, arg = "data")
}
\arguments{
\item{x}{The object to check.}

\item{arg}{Argument name for the error message.}
}
\value{
\code{x}, invisibly.
}
\description{
Check that an object is a SummarizedExperiment
}
\author{
Jared Andrews
}
\keyword{internal}
