% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_readers.R
\name{.assert_structure}
\alias{.assert_structure}
\title{Check that an object is from read_structure()}
\usage{
.assert_structure(x, arg = "data")
}
\arguments{
\item{x}{The object to check.}

\item{arg}{Argument name for the error message.}
}
\value{
\code{x}, invisibly.
}
\description{
Check that an object is from read_structure()
}
\author{
Jared Andrews
}
\keyword{internal}
