% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoDotPlot_module_ui.R
\name{.ddp_summary_fxn}
\alias{.ddp_summary_fxn}
\title{The colour summary function for dittoDotPlot}
\usage{
.ddp_summary_fxn(name)
}
\arguments{
\item{name}{\code{"nonzero.mean"}, \code{"mean"} or \code{"median"}.}
}
\value{
A function of a numeric vector.
}
\description{
The colour summary function for dittoDotPlot
}
\author{
Jared Andrews
}
\keyword{internal}
