% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoHeatmap_helpers.R
\name{.ditto_assays}
\alias{.ditto_assays}
\title{Assays a dittoSeq object offers for a heatmap}
\usage{
.ditto_assays(object)
}
\arguments{
\item{object}{A dittoSeq-compatible object.}
}
\value{
The assay names of a \code{SummarizedExperiment}-family object, or
\code{character(0)} (dittoSeq then picks the default assay itself).
}
\description{
Assays a dittoSeq object offers for a heatmap
}
\author{
Jared Andrews
}
\keyword{internal}
