% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoHeatmap_helpers.R
\name{.ditto_default_assay}
\alias{.ditto_default_assay}
\title{The assay dittoSeq uses by default}
\usage{
.ditto_default_assay(object)
}
\arguments{
\item{object}{A dittoSeq-compatible object.}
}
\value{
The first of \code{logcounts}, \code{normcounts} and \code{counts} present (as
dittoSeq chooses), otherwise the first assay, or \code{""}.
}
\description{
The assay dittoSeq uses by default
}
\author{
Jared Andrews
}
\keyword{internal}
