% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ditto_module_app.R
\name{.ditto_module_app}
\alias{.ditto_module_app}
\title{Build a standalone Shiny app for a dittoSeq-based module}
\usage{
.ditto_module_app(inputs_ui_fn, output_ui_fn, server_fn, object_list, title)
}
\arguments{
\item{inputs_ui_fn}{A module \verb{*InputsUI(id, data, ...)} function.}

\item{output_ui_fn}{A module \verb{*OutputUI(id)} function.}

\item{server_fn}{A module \verb{*Server(id, data, ...)} function.}

\item{object_list}{A named list of dittoSeq-compatible objects.}

\item{title}{A character string used as the page title.}
}
\value{
A \code{\link[shiny:shinyApp]{shiny::shinyApp()}} object.
}
\description{
Internal factory that mirrors \code{\link[VizModules:createModuleApp]{VizModules::createModuleApp()}} but for modules
whose data are \code{SingleCellExperiment}, \code{Seurat}, or \code{SummarizedExperiment}
objects rather than data frames. Because those objects cannot be edited in the
interactive data-table used by \code{\link[VizModules:createModuleApp]{VizModules::createModuleApp()}}, this factory
instead offers an object selector and a read-only metadata (\code{colData}) preview.
}
\author{
Jacob Martin
}
\keyword{internal}
