% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ditto_helpers.R
\name{.ditto_reductions}
\alias{.ditto_reductions}
\title{Safely list dimensionality reduction names in a dittoSeq object}
\usage{
.ditto_reductions(object)
}
\arguments{
\item{object}{A dittoSeq-compatible object.}
}
\value{
A character vector of reduction names, or \code{character(0)}.
}
\description{
Safely list dimensionality reduction names in a dittoSeq object
}
\author{
Jacob Martin, Jared Andrews
}
\keyword{internal}
