% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseResponse_module_server.R
\name{.dose_response_hidden_inputs}
\alias{.dose_response_hidden_inputs}
\title{Inputs hidden by the dose-response module}
\usage{
.dose_response_hidden_inputs()
}
\value{
A character vector of scatter plot input IDs to hide.
}
\description{
Hides colour/shape/split mappings and split-related point options that do
not apply to a single dose-response curve.
}
\author{
Jacob Martin
}
\keyword{internal}
