% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseResponse_module_server.R
\name{.dose_response_hidden_tabs}
\alias{.dose_response_hidden_tabs}
\title{Tabs hidden by the dose-response module}
\usage{
.dose_response_hidden_tabs()
}
\value{
A character vector of scatter plot tab names to hide.
}
\description{
Tabs hidden by the dose-response module
}
\author{
Jacob Martin
}
\keyword{internal}
