% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestPlot.R
\name{.forest_axis_label}
\alias{.forest_axis_label}
\title{Axis label for a forest plot's estimates}
\usage{
.forest_axis_label(model, conf.level)
}
\arguments{
\item{model}{The model type.}

\item{conf.level}{The confidence level.}
}
\value{
A character scalar.
}
\description{
Axis label for a forest plot's estimates
}
\author{
Jared Andrews
}
\keyword{internal}
