% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestPlot.R
\name{.forest_binary_outcome}
\alias{.forest_binary_outcome}
\title{Coerce a binary outcome to 0/1}
\usage{
.forest_binary_outcome(x, name)
}
\arguments{
\item{x}{The outcome column.}

\item{name}{The column's name, for the error message.}
}
\value{
An integer vector of 0/1 (and \code{NA}).
}
\description{
Coerce a binary outcome to 0/1
}
\author{
Jared Andrews
}
\keyword{internal}
