% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ditto_helpers.R
\name{get_default_reduction}
\alias{get_default_reduction}
\title{Choose a default dimensional reduction for a dittoSeq object}
\usage{
get_default_reduction(object)
}
\arguments{
\item{object}{A dittoSeq-compatible object.}
}
\value{
A length-one character naming a reduction, or \code{""} when none exist.
}
\description{
Returns the name of the reduction to use by default: the first available
reduction on the object, or an empty string when none are present.
}
\author{
Jacob Martin, Jared Andrews
}
\keyword{internal}
