% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_helpers.R
\name{.gsea_default_pathways}
\alias{.gsea_default_pathways}
\title{Pathways to show by default}
\usage{
.gsea_default_pathways(x, n = 3)
}
\arguments{
\item{x}{A bundle from \code{\link[=.gsea_input]{.gsea_input()}}.}

\item{n}{How many.}
}
\value{
The \code{n} most significant pathways when results are available,
otherwise the first \code{n}.
}
\description{
Pathways to show by default
}
\author{
Jared Andrews
}
\keyword{internal}
