% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_helpers.R
\name{.gwas_column_inputs}
\alias{.gwas_column_inputs}
\title{Column selectors shared by the GWAS modules}
\usage{
.gwas_column_inputs(
  ns,
  data,
  d,
  which = c("chr.col", "bp.col", "p.col", "snp.col")
)
}
\arguments{
\item{ns}{Namespace function.}

\item{data}{The summary statistics.}

\item{d}{Resolved defaults.}

\item{which}{Which selectors to include.}
}
\value{
A \code{tagList}.
}
\description{
Column selectors shared by the GWAS modules
}
\author{
Jared Andrews
}
\keyword{internal}
