% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_helpers.R
\name{.gwas_lambda}
\alias{.gwas_lambda}
\title{Genomic inflation factor}
\usage{
.gwas_lambda(p)
}
\arguments{
\item{p}{Numeric p-values.}
}
\value{
A single number.
}
\description{
The median of the 1-df chi-squared statistics the p-values correspond to,
divided by the median of that distribution (lambda GC).
}
\author{
Jared Andrews
}
\keyword{internal}
