% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_helpers.R
\name{.gwas_qq_defaults}
\alias{.gwas_qq_defaults}
\title{Default inputs for the GWAS QQ plot module}
\usage{
.gwas_qq_defaults(data, defaults = NULL)
}
\arguments{
\item{data}{The summary statistics.}

\item{defaults}{A named list of user defaults, or \code{NULL}.}
}
\value{
A named list of defaults.
}
\description{
Default inputs for the GWAS QQ plot module
}
\author{
Jared Andrews
}
\keyword{internal}
