% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_helpers.R
\name{.gwas_require_columns}
\alias{.gwas_require_columns}
\title{Stop when the GWAS columns could not be found}
\usage{
.gwas_require_columns(d, need = c("chr.col", "bp.col", "p.col"))
}
\arguments{
\item{d}{Resolved defaults with \code{chr.col}, \code{bp.col}, \code{p.col}.}

\item{need}{Which of them are required.}
}
\value{
\code{NULL}, invisibly; errors otherwise.
}
\description{
Stop when the GWAS columns could not be found
}
\author{
Jared Andrews
}
\keyword{internal}
