% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ditto_helpers.R
\name{.is_ditto_object}
\alias{.is_ditto_object}
\title{Check whether an object is a dittoSeq-compatible object}
\usage{
.is_ditto_object(object)
}
\arguments{
\item{object}{An object to test.}
}
\value{
\code{TRUE} when \code{object} is a \code{SingleCellExperiment}, \code{SummarizedExperiment},
or \code{Seurat} object, otherwise \code{FALSE}.
}
\description{
Check whether an object is a dittoSeq-compatible object
}
\author{
Jared Andrews
}
\keyword{internal}
