% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalCurve_helpers.R
\name{.km_axis_title}
\alias{.km_axis_title}
\title{Y-axis title of a survival curve}
\usage{
.km_axis_title(fun = NULL)
}
\arguments{
\item{fun}{The curve transformation, as for \code{\link[=.km_frame]{.km_frame()}}.}
}
\value{
A string.
}
\description{
Y-axis title of a survival curve
}
\author{
Jared Andrews
}
\keyword{internal}
