% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalCurve_helpers.R
\name{.km_pvalue_text}
\alias{.km_pvalue_text}
\title{Format a log-rank p-value for display}
\usage{
.km_pvalue_text(p)
}
\arguments{
\item{p}{The p-value.}
}
\value{
A string such as \code{"p = 0.0013"} or \code{"p < 0.0001"}.
}
\description{
Format a log-rank p-value for display
}
\author{
Jared Andrews
}
\keyword{internal}
