% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationLollipop_module_ui.R
\name{.lollipop_defaults}
\alias{.lollipop_defaults}
\title{Default inputs for the mutationLollipop module}
\usage{
.lollipop_defaults(m, defaults = NULL)
}
\arguments{
\item{m}{A result of \code{\link[=.maf_table]{.maf_table()}}.}

\item{defaults}{A named list of user defaults, or \code{NULL}.}
}
\value{
A named list of defaults.
}
\description{
Default inputs for the mutationLollipop module
}
\author{
Jared Andrews
}
\keyword{internal}
