% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf_helpers.R
\name{.maf_validate}
\alias{.maf_validate}
\title{Validate the data of a mutation module}
\usage{
.maf_validate(x)
}
\arguments{
\item{x}{The module's data.}
}
\value{
A result of \code{\link[=.maf_table]{.maf_table()}}.
}
\description{
The \code{validate} function the MAF modules pass to \code{\link[=.sci_plot_server]{.sci_plot_server()}}.
}
\author{
Jared Andrews
}
\keyword{internal}
