% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafSummary_module_ui.R
\name{.mafsum_defaults}
\alias{.mafsum_defaults}
\title{Default inputs for the mafSummary module}
\usage{
.mafsum_defaults(defaults = NULL)
}
\arguments{
\item{defaults}{A named list of user defaults, or \code{NULL}.}
}
\value{
A named list of defaults.
}
\description{
Default inputs for the mafSummary module
}
\author{
Jared Andrews
}
\keyword{internal}
