% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcatools_helpers.R
\name{.pca_components}
\alias{.pca_components}
\title{Component names of a PCAtools \code{pca} object}
\usage{
.pca_components(p)
}
\arguments{
\item{p}{A \code{pca} object.}
}
\value{
A character vector, e.g. \code{c("PC1", "PC2", ...)}.
}
\description{
Component names of a PCAtools \code{pca} object
}
\author{
Jared Andrews
}
\keyword{internal}
