% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcatools_helpers.R
\name{.pca_module_app}
\alias{.pca_module_app}
\title{Build a standalone Shiny app for a PCAtools module}
\usage{
.pca_module_app(inputs_ui_fn, output_ui_fn, server_fn, pca_list, title)
}
\arguments{
\item{inputs_ui_fn, output_ui_fn, server_fn}{The module's functions.}

\item{pca_list}{A named list of PCAtools \code{pca} objects.}

\item{title}{Page title.}
}
\value{
A \code{\link[shiny:shinyApp]{shiny::shinyApp()}} object.
}
\description{
\code{\link[=.sci_object_app]{.sci_object_app()}} over a named list of \code{pca} objects, previewing each
object's sample metadata.
}
\author{
Jared Andrews
}
\keyword{internal}
