% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlots.R
\name{.pca_retained_loadings}
\alias{.pca_retained_loadings}
\title{Variables retained by PCAtools' \code{rangeRetain} rule}
\usage{
.pca_retained_loadings(ld, range.retain)
}
\arguments{
\item{ld}{A data frame of loadings (variables x components).}

\item{range.retain}{Fraction of each component's loading range.}
}
\value{
Integer row indices of the retained variables, in first-retained order.
}
\description{
Variables retained by PCAtools' \code{rangeRetain} rule
}
\author{
Jared Andrews
}
\keyword{internal}
