% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_helpers.R
\name{.pk_columns}
\alias{.pk_columns}
\title{Detect the columns of concentration-time data}
\usage{
.pk_columns(data)
}
\arguments{
\item{data}{The data frame.}
}
\value{
A list with \code{subject}, \code{time}, \code{conc}, \code{dose}, \code{group} (column names or \code{""}).
}
\description{
Detect the columns of concentration-time data
}
\author{
Jared Andrews
}
\keyword{internal}
