% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkConcentrationTime_module_ui.R
\name{.pk_defaults}
\alias{.pk_defaults}
\title{Default inputs for the pkConcentrationTime module}
\usage{
.pk_defaults(data, defaults = NULL)
}
\arguments{
\item{data}{The concentration-time data frame.}

\item{defaults}{A named list of user defaults, or \code{NULL}.}
}
\value{
A named list of defaults.
}
\description{
Default inputs for the pkConcentrationTime module
}
\author{
Jared Andrews
}
\keyword{internal}
