% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkConcentrationTime_module_app.R
\name{.pk_example}
\alias{.pk_example}
\title{The example concentration-time data}
\usage{
.pk_example()
}
\value{
\link[datasets:Theoph]{datasets::Theoph} as a plain data frame, with \code{Subject} as character.
}
\description{
The example concentration-time data
}
\author{
Jared Andrews
}
\keyword{internal}
