% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_helpers.R
\name{.plate_zprime}
\alias{.plate_zprime}
\title{Z'-factor of an assay plate}
\usage{
.plate_zprime(pos, neg)
}
\arguments{
\item{pos, neg}{Readouts of the positive and negative control wells.}
}
\value{
The Z'-factor, or \code{NA} without at least two of each.
}
\description{
\verb{1 - 3 (sd_pos + sd_neg) / |mean_pos - mean_neg|} (Zhang et al. 1999).
}
\author{
Jared Andrews
}
\keyword{internal}
