% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goFanPlot.R
\name{.resolve_orgdb}
\alias{.resolve_orgdb}
\title{Resolve an OrgDb argument to an OrgDb object}
\usage{
.resolve_orgdb(org)
}
\arguments{
\item{org}{An \code{OrgDb} object or a length-one character naming an OrgDb
package.}
}
\value{
An \code{OrgDb} object.
}
\description{
Accepts either an \code{OrgDb} object (returned unchanged) or the name of an
installed organism annotation package (e.g. \code{"org.Hs.eg.db"}) and returns the
corresponding \code{OrgDb} object exported by that package.
}
\author{
Jacob Martin, Jared Andrews
}
\keyword{internal}
