% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_helpers.R
\name{.sci_plot_output_ui}
\alias{.sci_plot_output_ui}
\title{Plot output for a native plot module}
\usage{
.sci_plot_output_ui(id, name, resizable = TRUE, height = "400px")
}
\arguments{
\item{id}{The module id.}

\item{name}{The module name (the output id).}

\item{resizable}{Logical; wrap in \code{\link[shinyjqui:jqui_resizable]{shinyjqui::jqui_resizable()}}.}

\item{height}{CSS height of the output.}
}
\value{
The output UI.
}
\description{
Plot output for a native plot module
}
\author{
Jared Andrews
}
\keyword{internal}
