% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_heatmap_helpers.R
\name{.se_annotation_default}
\alias{.se_annotation_default}
\title{The first few sample-metadata columns that group samples (see \code{\link[=.se_group_cols]{.se_group_cols()}})}
\usage{
.se_annotation_default(meta, n = 2)
}
\arguments{
\item{meta}{A sample metadata data frame.}

\item{n}{The most columns to return.}
}
\value{
Column names.
}
\description{
The first few sample-metadata columns that group samples (see \code{\link[=.se_group_cols]{.se_group_cols()}})
}
\author{
Jared Andrews
}
\keyword{internal}
