% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_helpers.R
\name{.se_assays}
\alias{.se_assays}
\title{Assay names of a SummarizedExperiment}
\usage{
.se_assays(se)
}
\arguments{
\item{se}{A \code{SummarizedExperiment}.}
}
\value{
The assay names, or \code{"1"} for a single unnamed assay.
}
\description{
Assay names of a SummarizedExperiment
}
\author{
Jared Andrews
}
\keyword{internal}
