% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_helpers.R
\name{.se_default_assay}
\alias{.se_default_assay}
\title{The default assay of a SummarizedExperiment}
\usage{
.se_default_assay(se)
}
\arguments{
\item{se}{A \code{SummarizedExperiment}.}
}
\value{
\code{"counts"} when present, otherwise the first assay.
}
\description{
The default assay of a SummarizedExperiment
}
\author{
Jared Andrews
}
\keyword{internal}
