% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_helpers.R
\name{.se_label_cols}
\alias{.se_label_cols}
\title{Feature-label columns of a SummarizedExperiment}
\usage{
.se_label_cols(se)
}
\arguments{
\item{se}{A \code{SummarizedExperiment}.}
}
\value{
The names of the character or factor \code{rowData} columns.
}
\description{
Feature-label columns of a SummarizedExperiment
}
\author{
Jared Andrews
}
\keyword{internal}
