% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_helpers.R
\name{.se_module_app}
\alias{.se_module_app}
\title{Build a standalone Shiny app for a module that takes a SummarizedExperiment}
\usage{
.se_module_app(inputs_ui_fn, output_ui_fn, server_fn, object_list, title)
}
\arguments{
\item{inputs_ui_fn, output_ui_fn, server_fn}{The module's functions.}

\item{object_list}{A named list of \code{SummarizedExperiment} objects.}

\item{title}{Page title.}
}
\value{
A \code{\link[shiny:shinyApp]{shiny::shinyApp()}} object.
}
\description{
\code{\link[=.sci_object_app]{.sci_object_app()}} over a named list of experiments, previewing each one's
sample metadata.
}
\author{
Jared Andrews
}
\keyword{internal}
