% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_heatmap_helpers.R
\name{.se_stage}
\alias{.se_stage}
\title{Transform the assay of a SummarizedExperiment once, for the heatmap wrappers}
\usage{
.se_stage(se, assay, transform)
}
\arguments{
\item{se}{A \code{SummarizedExperiment}.}

\item{assay, transform}{As for \code{\link[=.se_transform]{.se_transform()}}.}
}
\value{
\code{list(se =, mat =)}, the object and its transformed assay.
}
\description{
Transform the assay of a SummarizedExperiment once, for the heatmap wrappers
}
\author{
Jared Andrews
}
\keyword{internal}
