% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_helpers.R
\name{.se_transform}
\alias{.se_transform}
\title{Transform a bulk expression assay}
\usage{
.se_transform(se, assay = NULL, method = c("vst", "log2cpm", "log2", "none"))
}
\arguments{
\item{se}{A \code{SummarizedExperiment}.}

\item{assay}{The assay to use.}

\item{method}{\code{"vst"} (\code{\link[DESeq2:vst]{DESeq2::vst()}}), \code{"log2cpm"} (as \code{\link[edgeR:cpm]{edgeR::cpm()}} with
\code{log = TRUE}), \code{"log2"} (\code{log2(x + 1)}) or \code{"none"}.}
}
\value{
A numeric matrix, features by samples.
}
\description{
Transform a bulk expression assay
}
\author{
Jared Andrews
}
\keyword{internal}
