% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureViewer_module_ui.R
\name{.structure_defaults}
\alias{.structure_defaults}
\title{Default inputs for the structureViewer module}
\usage{
.structure_defaults(x, defaults = NULL)
}
\arguments{
\item{x}{An object from \code{\link[=read_structure]{read_structure()}}.}

\item{defaults}{A named list of user defaults, or \code{NULL}.}
}
\value{
A named list of defaults.
}
\description{
Default inputs for the structureViewer module
}
\author{
Jared Andrews
}
\keyword{internal}
