% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafSummary_module_ui.R
\name{mafSummaryOutputUI}
\alias{mafSummaryOutputUI}
\title{Output UI components for the mafSummary module}
\usage{
mafSummaryOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; when \code{TRUE} (the default) the plot output can be resized by dragging.}
}
\value{
A Shiny plotlyOutput.
}
\description{
Output UI components for the mafSummary module
}
\examples{
mafSummaryOutputUI("summary")
}
\author{
Jared Andrews
}
