% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattanPlot_module_ui.R
\name{manhattanPlotOutputUI}
\alias{manhattanPlotOutputUI}
\title{Output UI components for the manhattanPlot module}
\usage{
manhattanPlotOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical, whether to wrap the output in a resizable container.}
}
\value{
A Shiny plotlyOutput for the Manhattan plot.
}
\description{
Output UI components for the manhattanPlot module
}
\examples{
manhattanPlotOutputUI("manhattan")
}
\author{
Jared Andrews
}
