% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdTrajectoryMetrics_module_ui.R
\name{mdTrajectoryMetricsOutputUI}
\alias{mdTrajectoryMetricsOutputUI}
\title{Output UI components for the mdTrajectoryMetrics module}
\usage{
mdTrajectoryMetricsOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; when \code{TRUE} (the default) the plot output can be resized by dragging.}
}
\value{
A Shiny plotlyOutput.
}
\description{
Output UI components for the mdTrajectoryMetrics module
}
\examples{
mdTrajectoryMetricsOutputUI("md")
}
\author{
Jared Andrews
}
