% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm_kinetics}
\alias{mm_kinetics}
\title{Example Michaelis-Menten enzyme-kinetics observations (simulated)}
\format{
A data frame with 19 rows and 2 columns:
\describe{
\item{S}{Substrate concentration (mM)}
\item{v}{Reaction velocity (dE/min)}
}
}
\source{
Simulated in \code{data-raw/generate_mm_kinetics.R}; no real
biological data are included.
}
\usage{
data(mm_kinetics)
}
\description{
A small, self-contained set of enzyme-kinetics observations (substrate
concentration versus reaction velocity) provided as ready-to-use example
data for the \code{\link[=michaelisMentenServer]{michaelisMentenServer()}} module and \code{\link[=michaelisMentenPlot]{michaelisMentenPlot()}}.
}
\examples{
library(sciVizModules)
data(mm_kinetics)
head(mm_kinetics)

}
\seealso{
\link{mm_kinetics_line},
\link{mm_kinetics_fit}
}
\author{
Jacob Martin
}
\keyword{datasets}
