% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationLollipop_module_ui.R
\name{mutationLollipopOutputUI}
\alias{mutationLollipopOutputUI}
\title{Output UI components for the mutationLollipop module}
\usage{
mutationLollipopOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; when \code{TRUE} (the default) the plot output can be resized by dragging.}
}
\value{
A Shiny plotlyOutput.
}
\description{
Output UI components for the mutationLollipop module
}
\examples{
mutationLollipopOutputUI("lollipop")
}
\author{
Jared Andrews
}
