% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoPlot_module_ui.R
\name{oncoPlotOutputUI}
\alias{oncoPlotOutputUI}
\title{Output UI components for the oncoPlot module}
\usage{
oncoPlotOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; when \code{TRUE} (the default) the plot output can be resized by dragging.}
}
\value{
A Shiny plotlyOutput.
}
\description{
Output UI components for the oncoPlot module
}
\examples{
oncoPlotOutputUI("onco")
}
\author{
Jared Andrews
}
