% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaBiplot_module_ui.R
\name{pcaBiplotOutputUI}
\alias{pcaBiplotOutputUI}
\title{Output UI components for the pcaBiplot module}
\usage{
pcaBiplotOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical, whether to wrap the output in a resizable container.}
}
\value{
A Shiny plotlyOutput for the biplot.
}
\description{
This should be placed in the UI where the plot should be shown.
}
\examples{
pcaBiplotOutputUI("biplot")
}
\author{
Jared Andrews
}
