% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaEigencorPlot_module_ui.R
\name{pcaEigencorPlotOutputUI}
\alias{pcaEigencorPlotOutputUI}
\title{Output UI components for the pcaEigencorPlot module}
\usage{
pcaEigencorPlotOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; when \code{TRUE} (the default) the plot output can be resized by dragging.}
}
\value{
A Shiny plotlyOutput for the correlation heatmap.
}
\description{
Output UI components for the pcaEigencorPlot module
}
\examples{
pcaEigencorPlotOutputUI("eigencor")
}
\author{
Jared Andrews
}
