% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaLoadingsPlot_module_ui.R
\name{pcaLoadingsPlotOutputUI}
\alias{pcaLoadingsPlotOutputUI}
\title{Output UI components for the pcaLoadingsPlot module}
\usage{
pcaLoadingsPlotOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; when \code{TRUE} (the default) the plot output can be resized by dragging.}
}
\value{
A Shiny plotlyOutput for the loadings plot.
}
\description{
Output UI components for the pcaLoadingsPlot module
}
\examples{
pcaLoadingsPlotOutputUI("loadings")
}
\author{
Jared Andrews
}
