% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPairsPlot_module_ui.R
\name{pcaPairsPlotOutputUI}
\alias{pcaPairsPlotOutputUI}
\title{Output UI components for the pcaPairsPlot module}
\usage{
pcaPairsPlotOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; when \code{TRUE} (the default) the plot output can be resized by dragging.}
}
\value{
A Shiny plotlyOutput for the pairs plot.
}
\description{
Output UI components for the pcaPairsPlot module
}
\examples{
pcaPairsPlotOutputUI("pairs")
}
\author{
Jared Andrews
}
