% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaScreePlot_module_ui.R
\name{pcaScreePlotOutputUI}
\alias{pcaScreePlotOutputUI}
\title{Output UI components for the pcaScreePlot module}
\usage{
pcaScreePlotOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; when \code{TRUE} (the default) the plot output can be resized by dragging.}
}
\value{
A Shiny plotlyOutput for the scree plot.
}
\description{
Output UI components for the pcaScreePlot module
}
\examples{
pcaScreePlotOutputUI("scree")
}
\author{
Jared Andrews
}
