% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkConcentrationTime_module_ui.R
\name{pkConcentrationTimeOutputUI}
\alias{pkConcentrationTimeOutputUI}
\title{Output UI components for the pkConcentrationTime module}
\usage{
pkConcentrationTimeOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; when \code{TRUE} (the default) the plot output can be resized by dragging.}
}
\value{
A Shiny plotlyOutput.
}
\description{
Output UI components for the pkConcentrationTime module
}
\examples{
pkConcentrationTimeOutputUI("pk")
}
\author{
Jared Andrews
}
