% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateHeatmap_module_ui.R
\name{plateHeatmapOutputUI}
\alias{plateHeatmapOutputUI}
\title{Output UI components for the plateHeatmap module}
\usage{
plateHeatmapOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; when \code{TRUE} (the default) the plot output can be resized by dragging.}
}
\value{
A Shiny plotlyOutput.
}
\description{
Output UI components for the plateHeatmap module
}
\examples{
plateHeatmapOutputUI("plate")
}
\author{
Jared Andrews
}
