% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocCurve_module_ui.R
\name{rocCurveOutputUI}
\alias{rocCurveOutputUI}
\title{Output UI components for the rocCurve module}
\usage{
rocCurveOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; when \code{TRUE} (the default) the plot output can be resized by dragging.}
}
\value{
A Shiny plotlyOutput.
}
\description{
Output UI components for the rocCurve module
}
\examples{
rocCurveOutputUI("roc")
}
\author{
Jared Andrews
}
