% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePCA_module_ui.R
\name{samplePCAOutputUI}
\alias{samplePCAOutputUI}
\title{Output UI components for the samplePCA module}
\usage{
samplePCAOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical, whether to wrap the output in a resizable container.}
}
\value{
A Shiny plotlyOutput for the PCA.
}
\description{
Output UI components for the samplePCA module
}
\examples{
samplePCAOutputUI("pca")
}
\author{
Jared Andrews
}
