% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureViewer_module_ui.R
\name{structureViewerOutputUI}
\alias{structureViewerOutputUI}
\title{Output UI components for the structureViewer module}
\usage{
structureViewerOutputUI(id, height = "600px")
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{height}{Height of the viewer, as a valid CSS unit.}
}
\value{
The viewer output.
}
\description{
Output UI components for the structureViewer module
}
\examples{
structureViewerOutputUI("viewer")
}
\author{
Jared Andrews
}
