% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnSegmentPlot_module_ui.R
\name{cnSegmentPlotOutputUI}
\alias{cnSegmentPlotOutputUI}
\title{Output UI components for the cnSegmentPlot module}
\usage{
cnSegmentPlotOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; when \code{TRUE} (the default) the plot output is
wrapped in \code{\link[shinyjqui:jqui_resizable]{shinyjqui::jqui_resizable()}} so it can be resized by dragging.}
}
\value{
A Shiny plotlyOutput for the copy number segment plot.
}
\description{
This should be placed in the UI where the plot should be shown.
}
\author{
Jared Andrews
}
