% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoBarPlot_module_ui.R
\name{dittoBarPlotInputsUI}
\alias{dittoBarPlotInputsUI}
\title{Input UI components for the dittoBarPlot module}
\usage{
dittoBarPlotInputsUI(id, data, defaults = NULL, title = "BarPlot Settings", columns = 2)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{data}{A \code{SingleCellExperiment}, \code{Seurat}, or
\code{SummarizedExperiment} object used to populate the input choices.}

\item{defaults}{A named list of default values for the inputs.}

\item{title}{An optional title for the UI grid.}

\item{columns}{Number of columns for the UI grid.}
}
\value{
A Shiny tagList containing the UI elements.
}
\description{
This should be placed in the UI where the inputs should be shown, with an \code{id}
that matches the \code{id} used in the \code{dittoBarPlotServer()} and
\code{dittoBarPlotOutputUI()} functions. The module renders a composition bar plot of a discrete variable per group (counts or proportions)
via \code{\link[dittoSeq:dittoBarPlot]{dittoSeq::dittoBarPlot()}}.
}
\seealso{
\code{\link[dittoSeq:dittoBarPlot]{dittoSeq::dittoBarPlot()}}, \code{\link[=dittoBarPlotOutputUI]{dittoBarPlotOutputUI()}}, \code{\link[=dittoBarPlotServer]{dittoBarPlotServer()}}, \code{\link[=dittoBarPlotApp]{dittoBarPlotApp()}}
}
\author{
Jared Andrews
}
