% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoDimPlot_module_ui.R
\name{dittoDimPlotInputsUI}
\alias{dittoDimPlotInputsUI}
\title{Input UI components for the dittoDimPlot module}
\usage{
dittoDimPlotInputsUI(id, data, defaults = NULL, title = "DimPlot Settings", columns = 2)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{data}{A \code{SingleCellExperiment}, \code{Seurat}, or
\code{SummarizedExperiment} object used to populate the input choices.}

\item{defaults}{A named list of default values for the inputs.}

\item{title}{An optional title for the UI grid.}

\item{columns}{Number of columns for the UI grid.}
}
\value{
A Shiny tagList containing the UI elements.
}
\description{
This should be placed in the UI where the inputs should be shown, with an \code{id}
that matches the \code{id} used in the \code{dittoDimPlotServer()} and
\code{dittoDimPlotOutputUI()} functions. The module renders a dimensionality-reduction embedding colored by a gene or metadata variable
via \code{\link[dittoSeq:dittoDimPlot]{dittoSeq::dittoDimPlot()}}.
}
\seealso{
\code{\link[dittoSeq:dittoDimPlot]{dittoSeq::dittoDimPlot()}}, \code{\link[=dittoDimPlotOutputUI]{dittoDimPlotOutputUI()}}, \code{\link[=dittoDimPlotServer]{dittoDimPlotServer()}}, \code{\link[=dittoDimPlotApp]{dittoDimPlotApp()}}
}
\author{
Jared Andrews
}
