% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoDimPlot_module_ui.R
\name{dittoDimPlotOutputUI}
\alias{dittoDimPlotOutputUI}
\title{Output UI components for the dittoDimPlot module}
\usage{
dittoDimPlotOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; when \code{TRUE} (the default) the plot output is
wrapped in \code{\link[shinyjqui:jqui_resizable]{shinyjqui::jqui_resizable()}} so it can be resized by dragging.}
}
\value{
A Shiny plotlyOutput for the dim plot.
}
\description{
This should be placed in the UI where the plot should be shown.
}
\examples{
dittoDimPlotOutputUI("plot")
}
\author{
Jacob Martin, Jared Andrews
}
