% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoDimPlot_module_ui.R
\name{dittoDimPlotOutputUI}
\alias{dittoDimPlotOutputUI}
\title{Output UI for the dittoDimPlot module}
\usage{
dittoDimPlotOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; whether the plot output should be resizable.}
}
\value{
A Shiny tagList containing the plotly output element.
}
\description{
This should be placed in the UI where the plot should be shown, with an \code{id}
that matches the \code{id} used in the \code{dittoDimPlotServer()} and
\code{dittoDimPlotInputsUI()} functions.
}
\seealso{
\code{\link[dittoSeq:dittoDimPlot]{dittoSeq::dittoDimPlot()}}, \code{\link[=dittoDimPlotInputsUI]{dittoDimPlotInputsUI()}}, \code{\link[=dittoDimPlotServer]{dittoDimPlotServer()}}, \code{\link[=dittoDimPlotApp]{dittoDimPlotApp()}}
}
\author{
Jared Andrews
}
