% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoDimPlot_module_server.R
\name{dittoDimPlotServer}
\alias{dittoDimPlotServer}
\title{Server logic for the dittoDimPlot module}
\usage{
dittoDimPlotServer(
  id,
  data,
  hide.inputs = NULL,
  hide.tabs = NULL,
  defaults = NULL
)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{data}{A \code{reactive} returning a \code{SingleCellExperiment}, \code{Seurat}, or
\code{SummarizedExperiment} object.}

\item{hide.inputs}{A character vector of input IDs to hide. These will still be
initialized and their values passed to the plot function, but the user will
not be able to see/adjust them in the UI.}

\item{hide.tabs}{A character vector of tab names to hide.}

\item{defaults}{A named list of default values used when resetting the inputs.
Typically the same list passed to \code{\link[=dittoDimPlotInputsUI]{dittoDimPlotInputsUI()}}.}
}
\value{
The \code{moduleServer} function for the dittoDimPlot module.
}
\description{
This module builds a dimensionality-reduction plot with
\code{\link[dittoSeq:dittoDimPlot]{dittoSeq::dittoDimPlot()}} and renders it as an interactive \code{plotly} figure.
}
\seealso{
\code{\link[dittoSeq:dittoDimPlot]{dittoSeq::dittoDimPlot()}}, \code{\link[=dittoDimPlotInputsUI]{dittoDimPlotInputsUI()}},
\code{\link[=dittoDimPlotOutputUI]{dittoDimPlotOutputUI()}}, \code{\link[=dittoDimPlotApp]{dittoDimPlotApp()}}
}
\author{
Jacob Martin, Jared Andrews
}
