% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoFreqPlot_module_ui.R
\name{dittoFreqPlotInputsUI}
\alias{dittoFreqPlotInputsUI}
\title{Input UI components for the dittoFreqPlot module}
\usage{
dittoFreqPlotInputsUI(id, data, defaults = NULL, title = "FreqPlot Settings", columns = 2)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{data}{A \code{SingleCellExperiment}, \code{Seurat}, or
\code{SummarizedExperiment} object used to populate the input choices.}

\item{defaults}{A named list of default values for the inputs.}

\item{title}{An optional title for the UI grid.}

\item{columns}{Number of columns for the UI grid.}
}
\value{
A Shiny tagList containing the UI elements.
}
\description{
This should be placed in the UI where the inputs should be shown, with an \code{id}
that matches the \code{id} used in the \code{dittoFreqPlotServer()} and
\code{dittoFreqPlotOutputUI()} functions. The module renders a per-sample frequency of a discrete variable across groups
via \code{\link[dittoSeq:dittoFreqPlot]{dittoSeq::dittoFreqPlot()}}.
}
\seealso{
\code{\link[dittoSeq:dittoFreqPlot]{dittoSeq::dittoFreqPlot()}}, \code{\link[=dittoFreqPlotOutputUI]{dittoFreqPlotOutputUI()}}, \code{\link[=dittoFreqPlotServer]{dittoFreqPlotServer()}}, \code{\link[=dittoFreqPlotApp]{dittoFreqPlotApp()}}
}
\author{
Jared Andrews
}
