% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoFreqPlot_module_ui.R
\name{dittoFreqPlotOutputUI}
\alias{dittoFreqPlotOutputUI}
\title{Output UI for the dittoFreqPlot module}
\usage{
dittoFreqPlotOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; whether the plot output should be resizable.}
}
\value{
A Shiny tagList containing the plotly output element.
}
\description{
This should be placed in the UI where the plot should be shown, with an \code{id}
that matches the \code{id} used in the \code{dittoFreqPlotServer()} and
\code{dittoFreqPlotInputsUI()} functions.
}
\seealso{
\code{\link[dittoSeq:dittoFreqPlot]{dittoSeq::dittoFreqPlot()}}, \code{\link[=dittoFreqPlotInputsUI]{dittoFreqPlotInputsUI()}}, \code{\link[=dittoFreqPlotServer]{dittoFreqPlotServer()}}, \code{\link[=dittoFreqPlotApp]{dittoFreqPlotApp()}}
}
\author{
Jared Andrews
}
