% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoPlot_module_ui.R
\name{dittoPlotOutputUI}
\alias{dittoPlotOutputUI}
\title{Output UI components for the dittoPlot module}
\usage{
dittoPlotOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; when \code{TRUE} (the default) the plot output is
wrapped in \code{\link[shinyjqui:jqui_resizable]{shinyjqui::jqui_resizable()}} so it can be resized by dragging.}
}
\value{
A Shiny plotlyOutput for the plot.
}
\description{
This should be placed in the UI where the plot should be shown.
}
\examples{
dittoPlotOutputUI("plot")
}
\author{
Jacob Martin, Jared Andrews
}
