% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoRidgeJitter_module_ui.R
\name{dittoRidgeJitterInputsUI}
\alias{dittoRidgeJitterInputsUI}
\title{Input UI components for the dittoRidgeJitter module}
\usage{
dittoRidgeJitterInputsUI(id, data, defaults = NULL, title = "RidgeJitter Settings", columns = 2)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{data}{A \code{SingleCellExperiment}, \code{Seurat}, or
\code{SummarizedExperiment} object used to populate the input choices.}

\item{defaults}{A named list of default values for the inputs.}

\item{title}{An optional title for the UI grid.}

\item{columns}{Number of columns for the UI grid.}
}
\value{
A Shiny tagList containing the UI elements.
}
\description{
This should be placed in the UI where the inputs should be shown, with an \code{id}
that matches the \code{id} used in the \code{dittoRidgeJitterServer()} and
\code{dittoRidgeJitterOutputUI()} functions. The module renders a ridgeline-with-jitter distribution plot
via \code{\link[dittoSeq:dittoRidgeJitter]{dittoSeq::dittoRidgeJitter()}}.
}
\seealso{
\code{\link[dittoSeq:dittoRidgeJitter]{dittoSeq::dittoRidgeJitter()}}, \code{\link[=dittoRidgeJitterOutputUI]{dittoRidgeJitterOutputUI()}}, \code{\link[=dittoRidgeJitterServer]{dittoRidgeJitterServer()}}, \code{\link[=dittoRidgeJitterApp]{dittoRidgeJitterApp()}}
}
\author{
Jared Andrews
}
