% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoRidgeJitter_module_ui.R
\name{dittoRidgeJitterOutputUI}
\alias{dittoRidgeJitterOutputUI}
\title{Output UI for the dittoRidgeJitter module}
\usage{
dittoRidgeJitterOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; whether the plot output should be resizable.}
}
\value{
A Shiny tagList containing the plotly output element.
}
\description{
This should be placed in the UI where the plot should be shown, with an \code{id}
that matches the \code{id} used in the \code{dittoRidgeJitterServer()} and
\code{dittoRidgeJitterInputsUI()} functions.
}
\seealso{
\code{\link[dittoSeq:dittoRidgeJitter]{dittoSeq::dittoRidgeJitter()}}, \code{\link[=dittoRidgeJitterInputsUI]{dittoRidgeJitterInputsUI()}}, \code{\link[=dittoRidgeJitterServer]{dittoRidgeJitterServer()}}, \code{\link[=dittoRidgeJitterApp]{dittoRidgeJitterApp()}}
}
\author{
Jared Andrews
}
