% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoScatterPlot_module_ui.R
\name{dittoScatterPlotInputsUI}
\alias{dittoScatterPlotInputsUI}
\title{Input UI components for the dittoScatterPlot module}
\usage{
dittoScatterPlotInputsUI(id, data, defaults = NULL, title = "ScatterPlot Settings", columns = 2)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{data}{A \code{SingleCellExperiment}, \code{Seurat}, or
\code{SummarizedExperiment} object used to populate the input choices.}

\item{defaults}{A named list of default values for the inputs.}

\item{title}{An optional title for the UI grid.}

\item{columns}{Number of columns for the UI grid.}
}
\value{
A Shiny tagList containing the UI elements.
}
\description{
This should be placed in the UI where the inputs should be shown, with an \code{id}
that matches the \code{id} used in the \code{dittoScatterPlotServer()} and
\code{dittoScatterPlotOutputUI()} functions. The module renders a scatter plot of any two genes/metadata with an optional color variable
via \code{\link[dittoSeq:dittoScatterPlot]{dittoSeq::dittoScatterPlot()}}.
}
\seealso{
\code{\link[dittoSeq:dittoScatterPlot]{dittoSeq::dittoScatterPlot()}}, \code{\link[=dittoScatterPlotOutputUI]{dittoScatterPlotOutputUI()}}, \code{\link[=dittoScatterPlotServer]{dittoScatterPlotServer()}}, \code{\link[=dittoScatterPlotApp]{dittoScatterPlotApp()}}
}
\author{
Jared Andrews
}
