% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dittoScatterPlot_module_ui.R
\name{dittoScatterPlotOutputUI}
\alias{dittoScatterPlotOutputUI}
\title{Output UI for the dittoScatterPlot module}
\usage{
dittoScatterPlotOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; whether the plot output should be resizable.}
}
\value{
A Shiny tagList containing the plotly output element.
}
\description{
This should be placed in the UI where the plot should be shown, with an \code{id}
that matches the \code{id} used in the \code{dittoScatterPlotServer()} and
\code{dittoScatterPlotInputsUI()} functions.
}
\seealso{
\code{\link[dittoSeq:dittoScatterPlot]{dittoSeq::dittoScatterPlot()}}, \code{\link[=dittoScatterPlotInputsUI]{dittoScatterPlotInputsUI()}}, \code{\link[=dittoScatterPlotServer]{dittoScatterPlotServer()}}, \code{\link[=dittoScatterPlotApp]{dittoScatterPlotApp()}}
}
\author{
Jared Andrews
}
