% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseResponse_module_ui.R
\name{doseResponseOutputUI}
\alias{doseResponseOutputUI}
\title{Output UI components for the dose-response module}
\usage{
doseResponseOutputUI(id, resizable = TRUE)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{resizable}{Logical; when \code{TRUE} (the default) the plot output is
resizable.}
}
\value{
A Shiny plot output for the dose-response curve.
}
\description{
Thin wrapper over \code{\link[VizModules:dittoViz_scatterPlotOutputUI]{VizModules::dittoViz_scatterPlotOutputUI()}}.
}
\examples{
doseResponseOutputUI("dose")
}
\author{
Jacob Martin
}
