% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ditto_helpers.R
\name{.ditto_continuous_metas}
\alias{.ditto_continuous_metas}
\title{Identify continuous metadata columns in a dittoSeq object}
\usage{
.ditto_continuous_metas(object, max.levels = 30)
}
\arguments{
\item{object}{A dittoSeq-compatible object.}

\item{max.levels}{Numeric columns with more than this many unique values are
treated as continuous.}
}
\value{
A character vector of continuous metadata names.
}
\description{
Identify continuous metadata columns in a dittoSeq object
}
\author{
Jacob Martin, Jared Andrews
}
\keyword{internal}
