% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ditto_helpers.R
\name{.ditto_default_reduction}
\alias{.ditto_default_reduction}
\title{Choose a sensible default dimensionality reduction}
\usage{
.ditto_default_reduction(object)
}
\arguments{
\item{object}{A dittoSeq-compatible object.}
}
\value{
The name of the best-guess reduction (priority UMAP > t-SNE > PCA),
or \code{NULL} when none exist.
}
\description{
Choose a sensible default dimensionality reduction
}
\author{
Jared Andrews
}
\keyword{internal}
