% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ditto_helpers.R
\name{.ditto_genes}
\alias{.ditto_genes}
\title{Safely list gene names in a dittoSeq object}
\usage{
.ditto_genes(object)
}
\arguments{
\item{object}{A dittoSeq-compatible object.}
}
\value{
A character vector of gene names, or \code{character(0)}.
}
\description{
Safely list gene names in a dittoSeq object
}
\author{
Jacob Martin, Jared Andrews
}
\keyword{internal}
