% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ditto_helpers.R
\name{.ditto_group_levels}
\alias{.ditto_group_levels}
\title{Discrete levels of a metadata column, for palette groups}
\usage{
.ditto_group_levels(object, meta)
}
\arguments{
\item{object}{A dittoSeq-compatible object.}

\item{meta}{The name of a metadata column.}
}
\value{
A character vector of levels, or \code{character(0)}.
}
\description{
Discrete levels of a metadata column, for palette groups
}
\author{
Jared Andrews
}
\keyword{internal}
