% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ditto_helpers.R
\name{.ditto_meta_values}
\alias{.ditto_meta_values}
\title{Fetch the values of a single metadata column}
\usage{
.ditto_meta_values(object, meta)
}
\arguments{
\item{object}{A dittoSeq-compatible object.}

\item{meta}{The name of a metadata column.}
}
\value{
The metadata values, or \code{NULL} when unavailable.
}
\description{
Fetch the values of a single metadata column
}
\author{
Jared Andrews
}
\keyword{internal}
