% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ditto_helpers.R
\name{.ditto_metas}
\alias{.ditto_metas}
\title{Safely list metadata column names in a dittoSeq object}
\usage{
.ditto_metas(object)
}
\arguments{
\item{object}{A dittoSeq-compatible object.}
}
\value{
A character vector of metadata names, or \code{character(0)}.
}
\description{
Safely list metadata column names in a dittoSeq object
}
\author{
Jacob Martin, Jared Andrews
}
\keyword{internal}
