% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goFanPlot_helpers.R
\name{.gofan_palettes}
\alias{.gofan_palettes}
\title{Built-in colour palettes for the GO sunburst plot}
\usage{
.gofan_palettes
}
\description{
Named vector of plotly continuous colorscales suitable for the \code{palette}
argument of \code{\link[=goFanPlot]{goFanPlot()}} and the "Palette" UI input. Names are display
labels; values are the colorscale identifiers passed to plotly.
}
\author{
Jacob Martin
}
\keyword{internal}
