% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maPlot_module_ui.R
\name{maPlotOutputUI}
\alias{maPlotOutputUI}
\title{Output UI components for the maPlot module}
\usage{
maPlotOutputUI(id)
}
\arguments{
\item{id}{The ID for the Shiny module.}
}
\value{
A Shiny plotlyOutput for the MA plot
}
\description{
This should be placed in the UI where the plot should be shown.
}
\examples{
maPlotOutputUI("plot")
}
\author{
Jared Andrews
}
