% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm_kinetics_fit}
\alias{mm_kinetics_fit}
\title{Fitted Michaelis-Menten model for the example kinetics data}
\format{
An object of class \code{nls}.
}
\source{
Fitted in \code{data-raw/generate_mm_kinetics.R}.
}
\usage{
mm_kinetics_fit
}
\description{
An \code{\link[stats:nls]{stats::nls()}} fit of \eqn{v = V_m S / (K + S)} to \link{mm_kinetics}. Its
coefficients (\code{K} and \code{Vm}) provide the Michaelis constant and maximum
velocity used for the K / Vmax annotations in the \code{\link[=michaelisMentenServer]{michaelisMentenServer()}}
module.
}
\examples{
library(sciVizModules)
data(mm_kinetics_fit)
coef(mm_kinetics_fit)

}
\seealso{
\link{mm_kinetics}, \link{mm_kinetics_line}
}
\author{
Jacob Martin
}
\keyword{datasets}
