% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm_kinetics_line}
\alias{mm_kinetics_line}
\title{Fitted Michaelis-Menten curve for the example kinetics data}
\format{
A data frame with 100 rows and 2 columns:
\describe{
\item{S}{Substrate concentration (mM)}
\item{v}{Predicted reaction velocity (dE/min)}
}
}
\source{
Predicted from \link{mm_kinetics_fit} in
\code{data-raw/generate_mm_kinetics.R}.
}
\usage{
data(mm_kinetics_line)
}
\description{
The fitted Michaelis-Menten curve for \link{mm_kinetics}, predicted over a fine
grid of substrate concentrations. This is the pre-computed line (\code{mml})
passed as the \code{model} argument to \code{\link[=michaelisMentenPlot]{michaelisMentenPlot()}}.
}
\examples{
library(sciVizModules)
data(mm_kinetics_line)
head(mm_kinetics_line)

}
\seealso{
\link{mm_kinetics}, \link{mm_kinetics_fit}
}
\author{
Jacob Martin
}
\keyword{datasets}
