% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalCurve_module_server.R
\name{survivalCurveServer}
\alias{survivalCurveServer}
\title{Server logic for survivalCurve module}
\usage{
survivalCurveServer(
  id,
  data,
  hide.inputs = NULL,
  hide.tabs = NULL,
  defaults = NULL
)
}
\arguments{
\item{id}{The ID for the Shiny module.}

\item{data}{A \code{reactive} containing the data frame to plot. Must contain a
numeric follow-up time column and an event/status column.}

\item{hide.inputs}{A character vector of input IDs to hide. These will still be
initialized and their values passed to the plot function, but the user will
not be able to see/adjust them in the UI.}

\item{hide.tabs}{A character vector of tab names to hide.}

\item{defaults}{A named list of default values used when resetting the
inputs. Typically the same list passed to \code{\link[=survivalCurveInputsUI]{survivalCurveInputsUI()}}.}
}
\value{
The \code{moduleServer} function for the survivalCurve module.
}
\description{
This module fits a Kaplan-Meier survival curve with the
\href{https://cran.r-project.org/package=survminer}{survminer} package and renders
it as an interactive \code{plotly} figure via \code{\link[=survivalCurve]{survivalCurve()}}.
}
\seealso{
\code{\link[survminer:ggsurvplot]{survminer::ggsurvplot()}}, \code{\link[=survivalCurve]{survivalCurve()}},
\code{\link[=survivalCurveInputsUI]{survivalCurveInputsUI()}}, \code{\link[=survivalCurveOutputUI]{survivalCurveOutputUI()}},
\code{\link[=survivalCurveApp]{survivalCurveApp()}}
}
\author{
Jacob Martin
}
