% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot_module_ui.R
\name{volcanoPlotOutputUI}
\alias{volcanoPlotOutputUI}
\title{Output UI components for the volcanoPlot module}
\usage{
volcanoPlotOutputUI(id)
}
\arguments{
\item{id}{The ID for the Shiny module.}
}
\value{
A Shiny plotlyOutput for the volcano plot
}
\description{
This should be placed in the UI where the plot should be shown.
}
\author{
Jared Andrews
}
